Add CPU frequency pinning functionality
This is not just a hack. It also relies on "sudo sh -c" working
and it is tied to finding the right structure of /sys. And generally
not portable. But it works for me. And that is currently reason
enough for me to support it. Patches to improve portability and
robustness (yuck, how I am doing sudo) would be most welcome.