Refer to "github.com/tsee/dumbbench"
While I have no particularly strong feeling of ownership
over such things, in reality, brian has full access to the above
repo, and it has been the original repo and still has an
outstanding issue in the issue tracker. Thus is seems to
make more sense to standardize on using that repo as the
master where sharing links.